PARCIAL/requirements.txt
streamlit>=1.33.0
tensorflow>=2.12.0
numpy>=1.23.0
pillow>=9.0.0
